Delete old indexer
